+++ src/components/Dropzones/Dropzone.js
@@ -1,16 +1,13 @@
 import { useDrop } from 'react-dnd';
-import useStore from '../../lib/useStore';
 
 import SvgIcon from '../SVGs/icons';
 import { Dropzone } from '../UI/Dropzone/Dropzone.styled';
 import { ItemTypes } from '../UI/items';
 
-export default function Dropzone1(props) {
-	const moveOneUp = useStore(state => state.moveOneUp);
-
+export default function Dropzone1({ children, onMove }) {
 	const [{ isOver }, drop] = useDrop({
 		accept: ItemTypes.CARD,
-		drop: item => moveOneUp(item.index),
+		drop: item => onMove(item.index),
 		collect: monitor => ({
 			isOver: !!monitor.isOver(),
 		}),
@@ -18,7 +15,7 @@ export default function Dropzone1(props) {
 
 	return (
 		<Dropzone ref={drop} background={isOver ? '#d5c3dd' : '#f6f6f6'}>
-			{props.children}
+			{children}
 			<SvgIcon variant="dropzone" size="50px" color="#6667ab" />
 		</Dropzone>
 	);
