+++ src/components/Todos/GenericGrid.js
@@ -1,0 +1,41 @@
+import { Container } from '../UI/Grid/Container';
+import { ListHeadline } from '../UI/TodoCard/ListHeadline.styled';
+import DropWrapper from './DropWrapper';
+import { useDrop } from 'react-dnd';
+import { ItemTypes } from '../UI/items';
+import { TodoList } from '../UI/TodoCard/TodoList.styled';
+
+export default function GenericGrid({ todoList, dropzone, TodoComponent }) {
+	const [{ isOver }, drop] = useDrop({
+		accept: ItemTypes.CARD,
+
+		collect: monitor => ({
+			isOver: !!monitor.isOver(),
+		}),
+	});
+
+	return (
+		<Container>
+			{dropzone}
+			<ListHeadline>DoIT</ListHeadline>
+			<DropWrapper>
+				<TodoList
+					ref={drop}
+					role="list"
+					background={isOver ? 'rgba(178, 231, 159, 0.60)' : '#f6f6f6'}
+				>
+					{todoList.map((todo, index) => {
+						return (
+							<TodoComponent
+								key={todo.id}
+								id={todo.id}
+								title={todo.title}
+								index={index}
+							/>
+						);
+					})}
+				</TodoList>
+			</DropWrapper>
+		</Container>
+	);
+}
