+++ src/components/Todos/TodosGrid.js
@@ -1,38 +1,18 @@
-import { TodoList } from '../UI/TodoCard/TodoList.styled';
-import { Container } from '../UI/Grid/Container';
-import Todo from './Todo';
-import { ListHeadline } from '../UI/TodoCard/ListHeadline.styled';
-import Dropzone1 from '../Dropzones/Dropzone1';
+import Dropzone from '../Dropzones/Dropzone';
 import useStore from '../../lib/useStore';
-import DropWrapper from './DropWrapper';
-import { useDrop } from 'react-dnd';
-import { ItemTypes } from '../UI/items';
+
+import GenericGrid from './GenericGrid';
+import Todo from './Todo';
 
 export default function TodoGrid() {
 	const todos = useStore(state => state.todos);
-	const [{ isOver }, drop] = useDrop({
-		accept: ItemTypes.CARD,
-
-		collect: monitor => ({
-			isOver: !!monitor.isOver(),
-		}),
-	});
+	const moveOneUp = useStore(state => state.moveOneUp);
 
 	return (
-		<Container>
-			<Dropzone1 />
-			<ListHeadline>DoIT</ListHeadline>
-			<DropWrapper>
-				<TodoList
-					ref={drop}
-					role="list"
-					background={isOver ? 'rgba(178, 231, 159, 0.60)' : '#f6f6f6'}
-				>
-					{todos.map((todo, index) => {
-						return <Todo key={todo.id} id={todo.id} title={todo.title} index={index} />;
-					})}
-				</TodoList>
-			</DropWrapper>
-		</Container>
+		<GenericGrid
+			todoList={todos}
+			dropzone={<Dropzone onMove={moveOneUp} />}
+			TodoComponent={Todo}
+		/>
 	);
 }
